cli/requirements.txt
requests
web3
click
web3[tester]
pytest
mypy
python-dotenv